retail-demo/python/requirements.txt
azure-cosmos==4.13.0b2
azure-identity
aiohttp